login de medicos funcionando
correccion de modelo medico

solucion de errores en jwt inteceptor y error interceptor

+++ src/app/components/medic/medic-profile/medic-profile.component.ts
@@ -39,10 +39,12 @@ export class MedicProfileComponent implements OnInit {
   }
 
   deleteMedic(id: string) {
-    this.medicService.delete(id)
-      .subscribe(res => {
-        console.log(res);
-        this.router.navigate(['/medic']);
-      });
+    if (window.confirm('Esta seguro?')) {
+      this.medicService.delete(id)
+        .subscribe(res => {
+          console.log(res);
+          this.router.navigate(['/medic']);
+        });
+      }
   }
 }
